+++ src/Components/ActiveView/ActiveView.css
@@ -1,0 +1,19 @@
+.openMenu {
+    position: fixed;
+    min-width: 50vw;
+    max-width: 100vw;
+    height: calc(80vh - 2px);
+    background: 'gray';
+    z-index: 10;
+    border-right: '1px solid black';
+    border-radius: '0px 10px 10px 0px';
+}
+
+.closedMenu {
+    position: fixed;
+    width: 0px;
+    height: 80vh;
+    overflow: hidden;
+    left: 0px;
+    z-index: 10
+}
