+++ src/Components/Header/Header.css
@@ -1,0 +1,79 @@
+.header {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    height: 10vh;
+    border-bottom: 1px solid black;
+    background: #C4C4C4;
+}
+
+.menuButton {
+    margin-left: 25px;
+    display: flex;
+    flex-direction: column;
+    background: white;
+}
+
+.profileButton {
+    margin-right: 25px;
+    background: white;
+}
+
+.profileButtonLoggedIn {
+    margin-right: 25px;
+    background: '#548F25';
+}
+
+.menuContainer {
+    display: flex;
+    flex-direction: column;
+    height: 50;
+    align-items: center;
+    justify-content: center;
+    margin-top: -5px;
+}
+
+.menuRow {
+    height: 7px;
+    width: 50px;
+    margin-top: 5px;
+    background: black;
+}
+
+.menuRowOpenA {
+    height: 7px;
+    width: 50px;
+    margin-top: 5px;
+    transform: rotate(45deg);
+    background: black;
+}
+
+.menuRowOpenB {
+    height: 7px;
+    width: 50px;
+    margin-top: -7.5px;
+    transform: rotate(135deg);
+    background: black;
+}
+
+.buttonImage {
+    width: 40;
+    height: 40;
+}
+
+.buttonImageClicked {
+    width: 35px;
+    height: 35px;
+}
+
+.buttonGeneral {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 60px;
+    height: 60px;
+    border-radius: 15px;
+    cursor: pointer;
+    border: 1px solid black;
+}
