+++ src/Components/LoginRegistration/LoginRegistration.css
@@ -1,0 +1,54 @@
+.loginComponent {
+    display: flex;
+    height: 80vh;
+}
+
+.loginContainer {
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 5vh;
+    height: 40vh;
+    width: 80vw;
+    max-width: 800px;
+    border-radius: 15px;
+    background: gray;
+}
+
+.textInput {
+    height: 25px;
+    width: 40vw;
+    border-radius: 10px;
+    margin-bottom: 10px;
+    text-align: center
+}
+
+.loginButton {
+    height: 25px;
+    width: 40vw;
+    margin-top: 41px;
+    margin-left: auto;
+    margin-right: auto;
+    border: 1px solid black;
+    border-radius: 10px;
+    cursor: pointer;
+    user-select: none;
+    background: #548F25
+}
+
+.registerButton {
+    height: 25px;
+    width: 40vw;
+    margin-left: auto;
+    margin-right: auto;
+    border: 1px solid black;
+    border-radius: 10px;
+    cursor: pointer;
+    user-select: none;
+    background: #548F25
+}
+
+.newCustomerText {
+    margin-top: 20px;
+    cursor: pointer;
+    user-select: none
+}
