+++ src/Components/LoginRegistration/LoginRegistration.jsx
@@ -1,7 +1,8 @@
 import { useState } from "react";
 import { useDispatch, useSelector } from 'react-redux';
 import firebase from 'firebase';
-import { actions } from '../Features/loggedinUser';
+import { actions } from '../../Features/loggedinUser';
+import './LoginRegistration.css';
 
 const LoginRegistration = () => {
     const [showLogin, setShowLogin] = useState(true);
@@ -13,59 +14,6 @@ const LoginRegistration = () => {
 
     const dispatch = useDispatch();
     console.log('loggedin: ', loggedin);
-
-    const style = {
-        loginComponent: {
-            display: 'flex',
-            height: '80vh',
-            // width: '100vw'
-        },
-        loginContainer: {
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            marginTop: '5vh',
-            height: '40vh',
-            width: '80vw',
-            maxWidth: '800px',
-            borderRadius: '15px',
-            background: 'gray'
-        },
-        textInput: {
-            height: '25px',
-            width: '40vw',
-            borderRadius: '10px',
-            marginBottom: '10px',
-            textAlign: 'center'
-        },
-        loginButton: {
-            height: '25px',
-            width: '40vw',
-            marginTop: '41px',
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            border: '1px solid black',
-            borderRadius: '10px',
-            cursor: 'pointer',
-            userSelect: 'none',
-            background: '#548F25'
-        },
-        registerButton: {
-            height: '25px',
-            width: '40vw',
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            border: '1px solid black',
-            borderRadius: '10px',
-            cursor: 'pointer',
-            userSelect: 'none',
-            background: '#548F25'
-        },
-        newCustomerText: {
-            marginTop: '20px',
-            cursor: 'pointer',
-            userSelect: 'none'
-        }
-    }
 
     function toggleLogin() {
         setUsername(userName);
@@ -115,26 +63,27 @@ const LoginRegistration = () => {
     }
 
     return (
-        <div style={style.loginComponent}>
+        // <div style={style.loginComponent}>
+        <div className="loginComponent">
             {showLogin ? (
-                <div style={style.loginContainer}>
+                <div className="loginContainer">
                     <h3>Login</h3>
-                    <input type='text' style={style.textInput} placeholder="E-mail" value={userName} onInput={e => setUsername(e.target.value)} />
-                    <input type='password' style={style.textInput} placeholder="Password" value={userPassword} onInput={e => setUserPassword(e.target.value)} />
-                    <div style={style.loginButton} onClick={() => loginUser()}>Login</div>
-                    <div style={style.newCustomerText} onClick={() => toggleLogin()}>
+                    <input type='text' className="textInput" placeholder="E-mail" value={userName} onInput={e => setUsername(e.target.value)} />
+                    <input type='password' className="textInput" placeholder="Password" value={userPassword} onInput={e => setUserPassword(e.target.value)} />
+                    <div className="loginButton" onClick={() => loginUser()}>Login</div>
+                    <div className="newCustomerText" onClick={() => toggleLogin()}>
                         Not yet a customer? <br />
                         Click here to sign up
                     </div>
                 </div>
             ) : (
-                <div style={style.loginContainer}>
+                <div className="loginContainer">
                     <h3>Registration</h3>
-                    <input type='text' style={style.textInput} placeholder="E-mail" value={userName} onInput={e => setUsername(e.target.value)} />
-                    <input type='password' style={style.textInput} placeholder="Password" value={userPassword} onInput={e => setUserPassword(e.target.value)} />
-                    <input type='password' style={style.textInput} placeholder="Confirm password" value={userPasswordConfirm} onInput={e => setUserPasswordConfirm(e.target.value)} />
-                    <div style={style.registerButton} onClick={() => registerUser()}>Register</div>
-                    <div style={style.newCustomerText} onClick={() => toggleLogin()}>
+                    <input type='text' className="textInput" placeholder="E-mail" value={userName} onInput={e => setUsername(e.target.value)} />
+                    <input type='password' className="textInput" placeholder="Password" value={userPassword} onInput={e => setUserPassword(e.target.value)} />
+                    <input type='password' className="textInput" placeholder="Confirm password" value={userPasswordConfirm} onInput={e => setUserPasswordConfirm(e.target.value)} />
+                    <div className="registerButton" onClick={() => registerUser()}>Register</div>
+                    <div className="newCustomerText" onClick={() => toggleLogin()}>
                         Back to login
                     </div>
                 </div>
